A lot of changes
All incoming messages will be added to queue,
if the message is designated to the receiver,
otherwise it is directly routed in the interrupt routine.

Get a message in the msg queue will get you the first
message that came in, not the last in queue.

Pinging parent is ongoing, disabled here, to make
the test cases pass.